Move all describer machinery out of `InspectableGraph`.
